requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.15
langchain_community==0.3.15
langchain_core==0.3.31
langchain_openai==0.3.2
langchain_text_splitters==0.3.5
langgraph==0.2.67
python-dotenv==1.0.1
streamlit==1.41.1
typing_extensions==4.12.2
